ai-assistant/requirements.txt
# AI 营销小助手 - 依赖
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# LLM 多厂商
openai==1.12.0
httpx==0.26.0

# 向量与文档解析
faiss-cpu
sentence-transformers
pypdf
python-pptx
openpyxl
python-docx

# 邮件
aiosmtplib==3.0.1

# 工具
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
